Add 32bits build (for dispmanx tests)

+++ src/DisplayPiImp.h
@@ -51,13 +51,13 @@ protected:
    CSpinLock   mutex_;
 
    DISPMANX_RESOURCE_HANDLE_T main_resource_;
-   unsigned int main_ptr_;
+   long unsigned int main_ptr_;
 
    DISPMANX_RESOURCE_HANDLE_T back_resource_;
-   unsigned int back_ptr_;
+   long unsigned int back_ptr_;
 
    DISPMANX_RESOURCE_HANDLE_T menu_resource_;
-   unsigned int menu_ptr_;
+   long unsigned int menu_ptr_;
 
    // Various display buffers
    int * display_buffer_[FRAME_BUFFER_SIZE];
